+++ Vice/app/src/main/java/martell/com/vice/activity/ArticleActivity.java
@@ -1,4 +1,4 @@
-package martell.com.vice;
+package martell.com.vice.activity;
 
 import android.content.Intent;
 import android.database.Cursor;
@@ -20,6 +20,7 @@ import com.facebook.share.widget.ShareButton;
 import com.nostra13.universalimageloader.core.ImageLoader;
 import com.nostra13.universalimageloader.core.ImageLoaderConfiguration;
 
+import martell.com.vice.R;
 import martell.com.vice.dbHelper.DatabaseHelper;
 import martell.com.vice.models.Article;
 import martell.com.vice.models.ArticleData;
