+++ Vice/app/src/main/java/martell/com/vice/adapters/ArticleAdapter.java
@@ -1,4 +1,4 @@
-package martell.com.vice;
+package martell.com.vice.adapters;
 
 import android.content.Context;
 import android.support.v7.widget.RecyclerView;
@@ -13,6 +13,7 @@ import com.squareup.picasso.Picasso;
 
 import java.util.List;
 
+import martell.com.vice.R;
 import martell.com.vice.models.Article;
 
 /**
