+++ Vice/app/src/main/java/martell/com/vice/nav_drawer/NotificationPublisher.java
@@ -1,15 +1,15 @@
-package martell.com.vice;
+package martell.com.vice.nav_drawer;
 
 import android.app.Notification;
 import android.app.NotificationManager;
 import android.app.PendingIntent;
-import android.app.TaskStackBuilder;
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
-import android.support.v7.app.NotificationCompat;
 import android.util.Log;
 
+import martell.com.vice.R;
+import martell.com.vice.activity.MainActivity;
 import martell.com.vice.dbHelper.NotificationDBHelper;
 
 /**
