+++ Vice/app/src/main/res/layout/activity_article.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="martell.com.vice.ArticleActivity">
+    tools:context=".activity.ArticleActivity">
     <android.support.design.widget.AppBarLayout
         android:id="@+id/appBar"
         android:layout_width="match_parent"
