+++ Vice/app/src/main/res/layout/activity_main.xml
@@ -8,7 +8,7 @@
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
     tools:openDrawer="start"
-    tools:context="martell.com.vice.MainActivity">
+    tools:context=".activity.MainActivity">
 
     <include
         layout="@layout/app_bar_main"
